refactor(dsh-mcp-manager): 阶段6批次3——connection/inject 域落位 + middleware-utils 拆分 + host-faces 收敛（Part of #664）
git mv 至目标目录：
- connection/orchestrator/：manager.ts（原根）
- connection/runtime/：supervisor/middleware/transport/protocol + middleware-const→limits
- inject/：middleware-register
- pipeline/authorize 并入策略裁决族（policyAllows/policyDenialReason/
  isToolDenied/toolDisabledReason）；parseDisabledTools 并入 config/store/
  middleware-state；middleware-utils.ts 删除
- types/host-faces.ts：四个最小面接口（ManagerLite/RoutesManager/
  MiddlewareHost/SupervisorLite）收敛（v3 §二），原定义处删除
- middleware.ts 六段 re-export 块删除（v3：汇聚只留 index.ts），
  index.ts 汇聚 re-export 全面重写（各域 interface.ts 拆分）
- connection/orchestrator/runtime/inject interface.ts 更新；verify-dir-imports
  符号存在性校验通过（74 文件、9 目录）

Design decisions:
- middleware.ts re-export 块删除是 v3 破环清单硬性项（汇聚收敛 index.ts），
  其符号面由 index.ts 从各域 interface.ts 直接取（lib 导出面零变化）
- RoutesManager 完整定义原样搬入 host-faces（含 sseHub/supervisors/
  catalogCache/middlewareMode/middleware 字段），零行为变更
- McpManager/McpMiddleware/McpMiddlewareHost 等类/接口沿 connection/
  interface.ts 导出（orchestrator 子层不对外）

Related issue #664
