+++ packages/dsh-mcp-manager/src/apply-runtime.ts
@@ -17,8 +17,8 @@ import type { PreStepDecision } from "@deepseek-ai/dsh-agent";
 import type { CatalogCache, CatalogDecision, CatalogMessage, SupervisorLite, CatalogAgent } from "./catalog/interface.ts";
 import { resolveCatalogInjection } from "./catalog/interface.ts";
 import { normalizeMiddlewareMode } from "./workspace/interface.ts";
-import type { McpManager } from "./manager.ts";
-import { registerMiddlewareTools, registerDirectMcpGuard } from "./middleware.ts";
+import type { McpManager } from "./connection/interface.ts";
+import { registerMiddlewareTools, registerDirectMcpGuard } from "./inject/interface.ts";
 import type { MiddlewareMode } from "./types/interface.ts";
 import { makeRoutes, makeEventsRoute, makeHealthRoute } from "./routes.ts";
 import { sseData } from "../../../shared/host-utils.js";
