+++ packages/dsh-mcp-manager/src/catalog/cache-view.ts
@@ -11,7 +11,7 @@
 import { existsSync, statSync } from "node:fs";
 import { join } from "node:path";
 import { dshHome } from "../../../../shared/dsh-home.js";
-import type { McpMiddleware } from "../middleware.ts";
+import type { McpMiddleware } from "../connection/interface.ts";
 import type { ServerConfig } from "../types/interface.ts";
 import { readCatalogServerFromDisk } from "../config/store/interface.ts";
 import { normalizedProjectRoot, SCOPE_PROJECT, MIDDLEWARE_GLOBAL_ROOT } from "../workspace/interface.ts";
