+++ packages/dsh-mcp-manager/src/config/model/config-schema.ts
@@ -8,7 +8,7 @@
 import z from "schemastery";
 import { clampZIndexBase, DEFAULT_Z_INDEX_BASE } from "../../placement-math.ts";
 import { DEFAULT_ANNOUNCE_CATALOG, DEFAULT_CATALOG_MAX_ENTRIES } from "../../catalog/interface.ts";
-import { DEFAULT_RESULT_TRUNCATE_BYTES } from "../../supervisor.ts";
+import { DEFAULT_RESULT_TRUNCATE_BYTES } from "../../connection/interface.ts";
 import type { ClientUiConfig, UiPlacementConfig } from "../../types/interface.ts";
 
 /** 空 description 工具的条件拼接默认开启。 */
