+++ packages/dsh-mcp-manager/src/config/model/normalize.ts
@@ -7,7 +7,7 @@
  */
 
 import type { ServerConfig } from "../../types/interface.ts";
-import { DEFAULT_TOOL_CALL_TIMEOUT_MS } from "../../supervisor.ts";
+import { DEFAULT_TOOL_CALL_TIMEOUT_MS } from "../../connection/interface.ts";
 
 /** MCP 服务器名命名空间约束（与官方 dsh-mcp-client 一致）。 */
 export const SERVER_NAME_PATTERN = /^[A-Za-z0-9_-]{1,32}$/;
