+++ packages/dsh-mcp-manager/src/config/store/interface.ts
@@ -13,5 +13,6 @@ export {
   readCatalogServerFromDisk,
   loadDisabledTools,
   saveDisabledTools,
+  parseDisabledTools,
 } from "./middleware-state.ts";
 export type { PersistedCatalogServer } from "./middleware-state.ts";
