+++ packages/dsh-mcp-manager/src/connection/runtime/protocol.ts
