+++ packages/dsh-mcp-manager/src/index.ts
@@ -61,9 +61,6 @@ export { resolveDebugConfig, resolveMiddlewareMode } from "./apply-config.ts";
 // 运行期装配工厂（组合根）：热切换为 B20/C-EVT 契约测试面
 export { makeMiddlewareHotSwitch } from "./apply-runtime.ts";
 
-// 服务器配置归一化（纯函数单一事实源）
-export { SERVER_NAME_PATTERN, normalizeServer } from "./config/model/interface.ts";
-
 // 插件 Config schema 与配置归一化（类型自 types.ts 取）
 export {
   DEFAULT_UI_CONFIG,
@@ -75,31 +72,8 @@ export {
 } from "./config/model/interface.ts";
 export type { UiPlacementConfig, ClientUiConfig } from "./types/interface.ts";
 
-// 管理器（MIDDLEWARE_GLOBAL_ROOT 单源在 workspace 域，阶段 4 收敛）
-export { McpManager } from "./manager.ts";
-export { MIDDLEWARE_GLOBAL_ROOT } from "./workspace/interface.ts";
-// 工作空间路由域（项目根发现 / 全名解析 / scope / 模式归一化；阶段 4 成形）
-export { findProjectRoot, normalizedProjectRoot, makeResolveRoot } from "./workspace/interface.ts";
-
-// 核心化 service（官方 storageDomain 模式）：ctx.mcpManager 类型面 + 声明合并。
-// 仅类型导出（无副作用导入）：消费方 import 类型时 tsc 会解析 service.d.ts，
-// 其内的 declare module 合并自动生效；副作用导入会让 stryker sandbox 解析
-// src/service.js 失败（sandbox 只有 .ts），也避免 .d.ts 里残留 .ts 引用。
-export type { McpManagerServerInput, McpManagerService } from "./integration/interface.ts";
-
-// 存储
-export { defaultStorePath, McpStore } from "./config/store/interface.ts";
-// 传输
-export { expandEnv, HttpTransport, parseSsePayload, StdioTransport, createTransport } from "./transport.ts";
-// 协议
-export { MCPClient } from "./protocol.ts";
-// CallToolResult 投影（协议结果 → 工具契约收敛，#512 单一事实源，归属 pipeline/project）
-export {
-  defaultCallResultFallbackText,
-  projectCallToolResult,
-} from "./pipeline/interface.ts";
-export type { CallResultTextHandlers, ProjectedCallResult } from "./pipeline/interface.ts";
-// 连接监督器 / 工具定义（含命名、截断、schema 校验）
+// 管理器 / 连接域（orchestrator+runtime：#664 阶段 6 集中搬移完成）
+export { McpManager, ConnectionSupervisor, McpMiddleware, expandEnv, HttpTransport, parseSsePayload, StdioTransport, createTransport, MCPClient } from "./connection/interface.ts";
 export {
   DEFAULT_TOOL_CALL_TIMEOUT_MS,
   DEFAULT_RESULT_TRUNCATE_BYTES,
@@ -107,11 +81,62 @@ export {
   truncateText,
   assertSupportedOutputSchema,
   buildToolDefinition,
-  ConnectionSupervisor,
-} from "./supervisor.ts";
-// 连接域（#664 阶段 3 收敛）：重连策略解析归 connection/runtime，经 connection/interface.ts 引用
-export { RECONNECT_DEFAULTS, resolveReconnect } from "./connection/interface.ts";
-// 能力目录 / 目录缓存（#664 阶段 5：catalog 域成形，经 catalog/interface.ts 转发）
+  RECONNECT_DEFAULTS,
+  resolveReconnect,
+  CONNECT_TIMEOUT_MS,
+  DISCOVERY_TIMEOUT_MS,
+  CALL_TIMEOUT_MS,
+  CATALOG_TTL_MS,
+  CATALOG_LRU_MAX,
+  MAX_TOOLS_PER_SERVER,
+  MAX_BYTES_PER_TOOL,
+  MAX_TOTAL_CATALOG_BYTES,
+  LIST_DEFAULT_TOOLS_PER_SERVER,
+  LIST_MAX_TOOLS_PER_SERVER,
+} from "./connection/interface.ts";
+export type { ReconnectPolicy } from "./connection/interface.ts";
+// 工作空间路由域（项目根发现 / 全名解析 / scope / 模式归一化；阶段 4 成形）
+export { findProjectRoot, normalizedProjectRoot, makeResolveRoot, MIDDLEWARE_GLOBAL_ROOT } from "./workspace/interface.ts";
+export {
+  fullServerName,
+  parseFullServerName,
+  normalizeToolName,
+  normalizeMiddlewareMode,
+} from "./workspace/interface.ts";
+// 执行管道域（两路径同构纯函数族；#664 阶段 2）
+export {
+  normalizeArguments,
+  msgOf,
+  createRedactor,
+  globMatch,
+  policyAllows,
+  policyDenialReason,
+  isToolDenied,
+  toolDisabledReason,
+  withTimeout,
+  defaultCallResultFallbackText,
+  projectCallToolResult,
+} from "./pipeline/interface.ts";
+export type { CallResultTextHandlers, ProjectedCallResult } from "./pipeline/interface.ts";
+// 核心化 service（官方 storageDomain 模式）：ctx.mcpManager 类型面 + 声明合并。
+// 仅类型导出（无副作用导入）：消费方 import 类型时 tsc 会解析 service.d.ts，
+// 其内的 declare module 合并自动生效；副作用导入会让 stryker sandbox 解析
+// src/service.js 失败（sandbox 只有 .ts），也避免 .d.ts 里残留 .ts 引用。
+export type { McpManagerServerInput, McpManagerService } from "./integration/interface.ts";
+
+// 存储与状态持久化（config/store：#664 阶段 6 落位）
+export { defaultStorePath, McpStore } from "./config/store/interface.ts";
+export {
+  userStateFile,
+  loadUserState,
+  saveUserState,
+  loadDisabledTools,
+  saveDisabledTools,
+  parseDisabledTools,
+  catalogCacheFileFor,
+  readCatalogServerFromDisk,
+} from "./config/store/interface.ts";
+// 能力目录 / 目录缓存（#664 阶段 5：catalog 域成形）
 export {
   DEFAULT_ANNOUNCE_CATALOG,
   DEFAULT_CATALOG_MAX_ENTRIES,
@@ -129,9 +154,16 @@ export {
   catalogHistory,
   renderMcpCatalogUpdate,
   resolveCatalogInjection,
+  scoreTool,
+  searchCatalog,
+  isCatalogFresh,
+  boundCatalogTools,
+  searchCatalogMulti,
+  listCatalog,
+  findToolDetail,
 } from "./catalog/interface.ts";
-// mcpServers JSON 导入
-export { fromClaudeEntry, parseClaudeJson } from "./config/model/interface.ts";
+// mcpServers JSON 导入 / 归一化（config/model）
+export { fromClaudeEntry, parseClaudeJson, SERVER_NAME_PATTERN, normalizeServer } from "./config/model/interface.ts";
 // 统计与 Debug
 export { McpStatsCollector, defaultStatsPath } from "./stats/interface.ts";
 export type {
@@ -141,51 +173,9 @@ export type {
   ProgressiveDisclosureStats,
   DebugConfig,
 } from "./stats/interface.ts";
-
-// 中间层（工作空间 MCP 路由：连接池 / 目录 / ws_mcp_search / ws_mcp_call /
-// ws_mcp_list / ws_mcp_detail）
-export {
-  McpMiddleware,
-  normalizeMiddlewareMode,
-  registerMiddlewareTools,
-  fullServerName,
-  parseFullServerName,
-  normalizeToolName,
-  normalizeArguments,
-  msgOf,
-  createRedactor,
-  globMatch,
-  policyAllows,
-  policyDenialReason,
-  isToolDenied,
-  toolDisabledReason,
-  parseDisabledTools,
-  scoreTool,
-  searchCatalog,
-  isCatalogFresh,
-  boundCatalogTools,
-  searchCatalogMulti,
-  listCatalog,
-  findToolDetail,
-  withTimeout,
-  userStateFile,
-  loadUserState,
-  saveUserState,
-  loadDisabledTools,
-  saveDisabledTools,
-  catalogCacheFileFor,
-  readCatalogServerFromDisk,
-  CONNECT_TIMEOUT_MS,
-  DISCOVERY_TIMEOUT_MS,
-  CALL_TIMEOUT_MS,
-  CATALOG_TTL_MS,
-  CATALOG_LRU_MAX,
-  MAX_TOOLS_PER_SERVER,
-  MAX_BYTES_PER_TOOL,
-  MAX_TOTAL_CATALOG_BYTES,
-  LIST_DEFAULT_TOOLS_PER_SERVER,
-  LIST_MAX_TOOLS_PER_SERVER,
-} from "./middleware.ts";
+// 工具注册面（inject：#664 阶段 6 落位）
+export { registerMiddlewareTools, registerDirectMcpGuard } from "./inject/interface.ts";
+// 共享类型面（types 域）
 export type {
   MiddlewareMode,
   MiddlewarePolicy,
@@ -196,7 +186,10 @@ export type {
   ListCatalogResult,
   ToolDetail,
   DisabledToolsMap,
-} from "./middleware.ts";
+  ServerConfig,
+  ServerStatus,
+} from "./types/interface.ts";
+
 // 路由
 export { ROUTES, makeRoutes, makeEventsRoute, makeHealthRoute, uiConfigChangedFrame, broadcastFrame, SSE_HEARTBEAT_MS, SSE_PING_FRAME } from "./routes.ts";
 export { SCOPE_GLOBAL, SCOPE_PROJECT, normalizeScope } from "./workspace/interface.ts";
