+++ packages/dsh-mcp-manager/src/workspace/root-resolution.ts
@@ -10,7 +10,7 @@
 import { existsSync } from "node:fs";
 import { dirname, join, resolve } from "node:path";
 import { dshHome } from "../../../../shared/dsh-home.js";
-import type { McpManager } from "../manager.ts";
+import type { McpManager } from "../connection/interface.ts";
 import { MIDDLEWARE_GLOBAL_ROOT } from "./constants.ts";
 
 /** DSH 全局家目录（shared/dsh-home.js 语义：DSH_HOME 非空白原样采用、空白
